+++ app/src/main/java/com/smart/camera/provider/RemoveProvider.java
@@ -13,6 +13,8 @@ import android.util.Log;
 import com.smart.camera.helper.DBOpenHelper;
 import com.smart.camera.tables.AIInfoTable;
 import com.smart.camera.tables.RemoveInfoTable;
+import com.smart.camera.tables.UploadInfoTable;
+import com.smart.camera.utils.DBUtil;
 
 import java.util.Objects;
 
@@ -60,8 +62,7 @@ public class RemoveProvider extends ContentProvider {
                 SQLiteDatabase db = dbOpenHelper.getReadableDatabase();
                 switch (MATCHER.match(uri)) {
                     case REMOVE_INFO_CODE:
-//                        return db.query(RemoveInfoTable.REMOVE_TABLE_NAME,projection,selection,selectionArgs,null,null,sortOrder);
-                        if(tableIsExist(RemoveInfoTable.REMOVE_TABLE_NAME)){
+                        if(removeTableIsExist(RemoveInfoTable.REMOVE_TABLE_NAME)){
                             return db.query(RemoveInfoTable.REMOVE_TABLE_NAME,projection,selection,selectionArgs,null,null,sortOrder);
                         }else {
                             return db.query(RemoveInfoTable.REMOVE_TABLE_NAME,null,null,null,null,null,null);
@@ -115,11 +116,18 @@ public class RemoveProvider extends ContentProvider {
     @Override
     public int delete(Uri uri, String selection, String[] selectionArgs) {
         SQLiteDatabase db = dbOpenHelper.getWritableDatabase();
+        //更新主键从1开始"
+        String sql = "update sqlite_sequence set seq=0 where name='" + RemoveInfoTable.REMOVE_TABLE_NAME + "'";
         int count = 0;
         switch (MATCHER.match(uri)) {
             case REMOVE_INFO_CODE:
-                count = db.delete(RemoveInfoTable.REMOVE_TABLE_NAME, selection, selectionArgs);
-                return count;
+                if(removeTableIsExist(RemoveInfoTable.REMOVE_TABLE_NAME)){
+                    count = db.delete(RemoveInfoTable.REMOVE_TABLE_NAME, selection, selectionArgs);
+                    db.execSQL(sql);
+                    return count;
+                }else {
+                    return count;
+                }
             default:
                 throw new IllegalArgumentException("Unkwon Uri:" + uri.toString());
         }
@@ -140,14 +148,18 @@ public class RemoveProvider extends ContentProvider {
         int count = 0;
         switch (MATCHER.match(uri)) {
             case REMOVE_INFO_CODE:
-                count = db.update(RemoveInfoTable.REMOVE_TABLE_NAME, values, selection, selectionArgs);
-                return count;
+                if(removeTableIsExist(RemoveInfoTable.REMOVE_TABLE_NAME)){
+                    count = db.update(RemoveInfoTable.REMOVE_TABLE_NAME, values, selection, selectionArgs);
+                    return count;
+                }else {
+                    return count;
+                }
             default:
                 throw new IllegalArgumentException("Unkwon Uri:" + uri.toString());
         }
     }
 
-    public boolean tableIsExist(String tableName){
+    public boolean removeTableIsExist(String tableName){
         boolean result = false;
         if(tableName == null){
             return false;
@@ -168,5 +180,4 @@ public class RemoveProvider extends ContentProvider {
         }
         return result;
     }
-
 }
